Feature date time tick (#11)
* feat: add datetime tick support

* feat: Optimized time scale rendering

+++ Plot.Core/Ticks/TickCollection.cs
@@ -4,23 +4,23 @@ namespace Plot.Core.Ticks
 {
     public class TickCollection
     {
-        public float[] Major { get; }
-        public float[] Minor { get; }
+        public double[] Major { get; }
+        public double[] Minor { get; }
         public string[] Labels { get; }
 
-        public TickCollection(float[] major, float[] minor, string[] labels)
+        public TickCollection(double[] major, double[] minor, string[] labels)
         {
             if (major.Length != labels.Length)
                 throw new InvalidOperationException($"{nameof(major)} must have the same length as {nameof(labels)}");
 
-            Major = major ?? Array.Empty<float>();
-            Minor = minor ?? Array.Empty<float>();
+            Major = major ?? Array.Empty<double>();
+            Minor = minor ?? Array.Empty<double>();
             Labels = labels ?? Array.Empty<string>();
         }
 
         public static TickCollection Empty => new TickCollection(
-            Array.Empty<float>(),
-            Array.Empty<float>(),
+            Array.Empty<double>(),
+            Array.Empty<double>(),
             Array.Empty<string>());
     }
 }
